requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
pydantic==2.5.3
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.2
pytz==2023.3
